requirements.txt
## Requirements.txt file

# Data Manipulation and Analysis
# numpy
pandas==2.0.3
python-dotenv==1.0.0

# # Data visualisation
# matplotlib
# seaborn

# # Cloud Services and Kubernetes
# kubernetes==26.1.0
prefect==2.11.3
# prefect-sqlalchemy==0.2.4
prefect-gcp[cloud_storage]==0.4.5
pandas_gbq
gcsfs

# # Database Interaction and ORM
# pgcli==3.5.0
# sqlalchemy==2.0.19

# # Web Requests and HTTP
# certifi==2023.7.22
# markdown==3.4.4
# requests==2.31.0
# requests-oauthlib==1.3.1

# # Configuration and Data Validation
# PyYAML==6.0.1
# pydantic==1.10.12

# # Concurrency and Asynchronous Programming
# anyio==3.7.1

# # Containerization and Virtualization
# docker==6.1.3


# Data Serialization and Formats
pyarrow==12.0.1
fastparquet==2023.7.0
# protobuf==4.23.4

# Jupyter Notebooks and IPython
ipykernel

# # Database Adapter
# psycopg2==2.9.6

# # Google BigQuery Integration
# pandas-gbq==0.19.2

# # Miscellaneous
# wget
# py4j




# # Data visualisation
# matplotlib==3.4.3
# seaborn==0.12.2
# statsmodels==0.14.0

# # Machine Learning
# scikit-learn==1.3.0

# # Filesystem handling 
# pathlib==1.0.1
# python-dotenv==1.0.0

# # Model Tracking
# mlflow==2.5.0

# # AWS SDK for Python - used for interacting with AWS services
# boto3==1.28.2

# # PostgreSQL 
# psycopg==3.1.9
# psycopg_binary==3.1.9
# psycopg2==2.9.6
# psycopg2_binary==2.9.6

# # Prefect
# prefect==2.10.21
# prefect-aws==0.3.5

# # Deployment
# flask==2.3.2
# gunicorn>=20,<22
# requests==2.28.1

# # Testing
# pytest==7.4.0
# pylint==2.17.4
# psutil==5.9.5